+++ packages/ws-controller/src/controller/MatterController.ts
@@ -390,6 +390,7 @@ export class MatterController {
                 this.#bleProxyEnabled,
                 !this.#disableOtaProvider,
                 this.#enableTimeSync,
+                !this.#threadDiagnosticsDisabled,
             );
 
             this.#commandHandler.events.started.once(async () => {